+++ src/parsley/field.js
@@ -39,7 +39,7 @@ define('parsley/field', [
         case 'pending': return null;
         case 'resolved': return true;
         case 'rejected': return this.validationResult;
-      };
+      }
     },
 
     // Validate field and trigger some events for mainly `ParsleyUI`
@@ -78,11 +78,15 @@ define('parsley/field', [
     // Just validate field. Do not trigger any event.
     // Returns `true` iff all constraints pass, `false` if there are failures,
     // or `null` if the result can not be determined yet (depends on a promise)
-    // Prefer using `whenValid`.
+    // See also `whenValid`.
     isValid: function (force, value) {
       return statusMapping[this.whenValid(force, value).state()];
     },
 
+    // Just validate field. Do not trigger any event.
+    // @returns a promise that succeeds only when all validations do.
+    // The argument `force` is optional, defaults to `false`.
+    // The argument `value` is optional. If given, it will be validated instead of the value of the input.
     whenValid: function (force, value) {
       // Recompute options and rebind constraints to have latest changes
       this.refreshConstraints();
@@ -92,6 +96,11 @@ define('parsley/field', [
       if (!this.hasConstraints())
         return $.when();
 
+      // Make `force` argument optional
+      if ('boolean' !== typeof force && 'undefined' === typeof value) {
+        value = force;
+        force = false;
+      }
       // Value could be passed as argument, needed to add more power to 'parsley:field:validate'
       if ('undefined' === typeof value || null === value)
         value = this.getValue();
@@ -291,8 +300,7 @@ define('parsley/field', [
     // Internal only.
     // Shortcut to trigger an event
     _trigger: function (eventName) {
-      eventName = 'field:' + eventName;
-      return this.trigger.apply(this, arguments);
+      return this.trigger('field:' + eventName);
     },
 
     // Internal only