+++ test/features/remote.js
@@ -183,6 +183,39 @@ define('features/remote', [
         expect(window.Parsley._remoteCache.dummy).to.be(undefined);
       });
 
+      it('should allow the change of XHR options', function (done) {
+        var parsleyInstance =
+          $('<input id="element" data-parsley-remote="http://parsleyjs.org" name="element" value="foobar"/>')
+          .appendTo('body')
+          .parsley()
+          .on('field:ajaxoptions', function (field, options) {
+            options.url = options.url + '/test/' + options.data.element;
+          });
+
+        stubAjax(200);
+        parsleyInstance.whenValid()
+          .done(function () {
+            expect($.ajax.calledWithMatch({ url: "http://parsleyjs.org/test/foobar" })).to.be(true);
+            expect($.ajax.calledWithMatch({ data: {element: 'foobar'} })).to.be(true);
+            done();
+          });
+      });
+
+      it('should allow RESTful URLs', function (done) {
+        var parsleyInstance =
+          $('<input id="element" data-parsley-remote="http://parsleyjs.org/thisisrest/{value}" name="element" value="foo bar"/>')
+          .appendTo('body')
+          .parsley();
+
+        stubAjax(200);
+        parsleyInstance.whenValid()
+          .done(function () {
+            expect($.ajax.calledWithMatch({ url: "http://parsleyjs.org/thisisrest/foo%20bar" })).to.be(true);
+            expect($.ajax.calledWithMatch({ data: {element: 'foo bar'} })).to.be(false);
+            done();
+          });
+      });
+
       it.skip('should abort successives querries and do not handle their return');
       afterEach(function () {
         $('#element, .parsley-errors-list').remove();