+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Parsley 2.x changelog
 
+## 2.2.0-rc2
+
+- Submit buttons: data now included in the submitted form (#826) and
+  attribute 'formnovalidate' is supported (#972)
+- Remote: use HTTP status code for what is considered valid or not (#956)
+- Remote: allow RESTful urls where "{value}" is replaced by the value to validate
+- Remote: add field:ajaxoptions to allow customizing of the ajax parameters (#894)
+- pattern validator is now anchored, unless it looks like /pattern/flag (#861)
+
 ## 2.2.0-rc1
 
 - Major validators refactor: